make changes and updating functions

+++ lib/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: tchaves <tchaves@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/19 15:17:42 by tchaves           #+#    #+#             */
-/*   Updated: 2023/10/20 14:30:55 by tchaves          ###   ########.fr       */
+/*   Updated: 2023/10/30 14:56:42 by tchaves          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@
 int	check_format(const char *arg, int i, va_list ap);
 int	ft_putchar(char c);
 int	ft_putunsnbr(unsigned int nb);
-int	ft_puthex(unsigned int d, int arg);
+int	ft_puthex(unsigned long d, int arg);
 int	ft_putstr(char *str);
 int	ft_putnbr(int nb);
